Construct models

+++ src/main/java/ai/quod/challenge/models/Repo.java
@@ -1,11 +1,11 @@
 package ai.quod.challenge.models;
 
 public class Repo {
-    private int id;
+    private long id;
     private String name;
     private String url;
 
-    public Repo(int id, String name, String url) {
+    public Repo(long id, String name, String url) {
         this.id = id;
         this.name = name;
         this.url = url;
